refactor(FOROME-722): small fixes
