+++ src/pages/filter/ui/modal-edit/components/modal-inheritance-mode/modal-edit-inheritance-mode.tsx
@@ -2,8 +2,8 @@ import { ReactElement, useEffect, useRef, useState } from 'react'
 import { observer } from 'mobx-react-lite'
 
 import dtreeStore from '@store/dtree'
+import { InheritanceModeContent } from '@pages/filter/ui/modal-edit/components/modal-inheritance-mode/components/inheritance-mode-content'
 import { HeaderModal } from '@pages/filter/ui/query-builder/ui/header-modal'
-import { InheritanceModeContent } from '@pages/filter/ui/query-builder/ui/inheritance-mode-content'
 import { ModalBase } from '@pages/filter/ui/query-builder/ui/modal-base'
 import modalEditStore from '../../modal-edit.store'
 import { EditModalButtons } from '../edit-modal-buttons'
